update the last 2 bugs in ReleaseServiceImpl and StoryServiceImpl

+++ src/main/java/com/smtteam/smt/service/impl/ReleaseServiceImpl.java
@@ -40,7 +40,10 @@ public class ReleaseServiceImpl implements ReleaseService {
 
     @Override
     public void deleteRelease(int id){
-        Release release = releaseDao.findById(id).get();
-        releaseDao.delete(release);
+        Optional<Release> optionalRelease = releaseDao.findById(id);
+        Release release = optionalRelease.orElse(null);
+        if( release != null){
+            releaseDao.delete(release);
+        }
     }
 }
